Populate charm docs with charm proxy and defer archive downloads
When the code invokes the AddCharmWithAuthorizationAndRepo method, the
refactored implementation will not download the full charm archive but
instead obtain a charm proxy (lightweight charm representation) by
invoking the charmstore API and use that to populate new charm docs in
the state.

The same method can be invoked a second time (via a worker) with the
ForceDownload parameter set to true to trigger the download and storing
of the charm archive in the controller's blobstore.