Implement retry mechanism for the manifest deployer
Given that charms are now downloaded asynchronously, the uniter might
receive a charm-not-found error when the agent attempts to fetch the
charm. This commit adds a retry mechanism (with exp backoff) to the
manifest deployer code that ensures that the agent will be able to
retrieve the charm it needs to install when the server has finished
downloading it.